+++ catalog/src/main/java/apap/tk/catalog/dto/request/UpdateCatalogDTO.java
@@ -5,13 +5,15 @@ import apap.tk.catalog.model.Category;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigInteger;
+
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class UpdateCatalogDTO {
     private String productName;
-    private int price;
+    private BigInteger price;
     private String productDescription;
     private int stock;
     private byte[] image;
